My tries to pivot my tables
I want to pivot the table in a way that i can also see how many cars from each model get each kind of complain but I find it hard to work around it right now as a newbie. I think I need to create a second pivot table for that? But then I will have to create two scatter plots to help me get an idea of what's happening with the data as well, right? 

Sorry for the many thoughts and questions; it just shows how messed up my thinking is right now.